requirements.txt
Package                Version
---------------------- -----------
camelot-py             0.9.0
cffi                   1.17.1
chardet                5.2.0
charset-normalizer     3.4.0
click                  8.1.7
cryptography           43.0.1
et-xmlfile             1.1.0
lxml                   5.3.0
mysql-connector-python 9.0.0
numpy                  2.1.2
opencv-python          4.10.0.84
openpyxl               3.1.5
packaging              24.1
pandas                 2.2.3
pdf2image              1.17.0
pdfminer.six           20231228
pdfplumber             0.11.4
pillow                 10.4.0
pip                    22.0.2
pycparser              2.22
PyMuPDF                1.24.11
PyPDF2                 2.10.6
pypdfium2              4.30.0
pytesseract            0.3.13
python-dateutil        2.9.0.post0
python-docx            1.1.2
python-pptx            1.0.2
pytz                   2024.2
setuptools             59.6.0
six                    1.16.0
typing_extensions      4.12.2
tzdata                 2024.2
XlsxWriter             3.2.0
